sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a7"
version = "0.2.2"
description = "A simple, synchronous Python SDK for the A7 Analytics Platform by Deutsche Börse"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Deutsche Börse", email = "support@deutsche-boerse.com" },
]
keywords = [
    "a7",
    "analytics",
    "deutsche-boerse",
    "eurex",
    "xetra",
    "market-data",
    "trading",
    "finance",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
    "ruff>=0.6.0",
    "pyright>=1.1.380",
    "build>=1.2.0",
    "twine>=5.1.0",
]

[project.urls]
Homepage = "https://github.com/Deutsche-Boerse/a7"
Documentation = "https://github.com/Deutsche-Boerse/a7/blob/master/sdk/README.md"
Repository = "https://github.com/Deutsche-Boerse/a7"
Issues = "https://github.com/Deutsche-Boerse/a7/issues"

[tool.hatch.build.targets.wheel]
packages = ["a7"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "integration: marks tests as integration tests (require A7_INTEGRATION_TESTS=1)",
]
# Register integration conftest as a plugin
pythonpath = ["tests"]

[tool.coverage.run]
source = ["a7"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 100
target-version = "py39"
src = ["a7", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",     # unused arguments (fixtures)
    "PLR2004", # magic values (test data)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pyright]
include = ["a7", "tests"]
exclude = [
    "**/__pycache__",
    "**/node_modules",
    ".venv",
]
pythonVersion = "3.9"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportPrivateUsage = false
